Fixed several bugs with generating entity lists.

+++ ConquestRTS/src/com/wraithavens/conquest/SinglePlayer/Entities/DynmapEntities/DictionaryEntry.java
@@ -3,10 +3,10 @@ package com.wraithavens.conquest.SinglePlayer.Entities.DynmapEntities;
 import com.wraithavens.conquest.SinglePlayer.Entities.EntityType;
 
 public class DictionaryEntry{
-	private final double spawnChance;
+	private final float spawnChance;
 	private final EntityType entity;
 	private final int variations;
-	DictionaryEntry(double spawnChance, EntityType entity, int variations){
+	DictionaryEntry(float spawnChance, EntityType entity, int variations){
 		this.spawnChance = spawnChance;
 		this.entity = entity;
 		this.variations = variations;
@@ -14,7 +14,7 @@ public class DictionaryEntry{
 	EntityType getEntity(){
 		return entity;
 	}
-	double getSpawnChance(){
+	float getSpawnChance(){
 		return spawnChance;
 	}
 	int getVariations(){
