+++ ConquestRTS/src/com/wraithavens/conquest/SinglePlayer/Entities/DynmapEntities/TempDictionaryEntry.java
@@ -3,15 +3,15 @@ package com.wraithavens.conquest.SinglePlayer.Entities.DynmapEntities;
 import com.wraithavens.conquest.SinglePlayer.Entities.EntityType;
 
 public class TempDictionaryEntry{
-	private final double averageDistance;
+	private final float averageDistance;
 	private final EntityType mainEntity;
 	private final int entityVariations;
-	TempDictionaryEntry(double averageDistance, EntityType mainEntity, int entityVariations){
+	TempDictionaryEntry(float averageDistance, EntityType mainEntity, int entityVariations){
 		this.averageDistance = averageDistance;
 		this.mainEntity = mainEntity;
 		this.entityVariations = entityVariations;
 	}
-	double getAverageDistance(){
+	float getAverageDistance(){
 		return averageDistance;
 	}
 	int getEntityVariations(){
